move internal inline function from .hxx to .cxx
This normally wouldn't make a difference, but it does when building
with Clang's -fmodules-codegen to collect duplicated templates,
inlines, etc. into extra .o file. The inline calling the template
is emitted in that .o, but the template is defined only in the .cxx
and not referenced from it, so it's never emitted.

Change-Id: I8b49030f2be6146ca1e2cd5e35734c0db6cca394
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87798
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>