DOCX export: fix table style config handling wrt nested tables
The bugdoc had 2 tables: both using the TableGrid table style, but one
had a direct formatting to disable all borders. The second was in the A1
cell of the first, and given that the table style config state was not
separated for nested tables, the border settings of the inner table
affected the settings of the later cells of the outer table.

Change-Id: Ie7897bc661d9f47ca9f5c1b3ed1c439ef0406037
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/87899
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins