DOCX export: fix handling of section starts that originally had headers
Both the DOC import (in wwSectionManager::InsertSegments()) and DOCX
import (in SectionPropertyMap::CloseSectionGroup()) have a mechanism to
try to attach changed headers/footers from a continuous section break
somewhere, so they are not lost.

This means that even if the rendering of such documents is OK, explicit
code is needed to undo the effect of the importer at export time, or
those headers will be lost.

Start doing this for the DOCX export case when the headers/footers are
placed at the "previous-in-practice" paragraph, more cases (handled at
the import side) can be added later.

Change-Id: Ic2304a74919d18da3ba9cb4afe301e0247a50dc2
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88103
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins