sw: DOCX export: fix rotated text fields
The rotation works in RTF export and DOC export, but not DOCX.

If a text field is rotated via SvxCharRotateItem, the w:eastAsianLayout
element gets written on the run that contains the fldChar begin,
but not the subsequent 4 runs that are written via
DoWriteFieldRunProperties(), because it doesn't write
m_pEastAsianLayoutAttrList.

It's not clear why that function manually writes a subset of the run
properties instead of simply calling WriteCollectedRunProperties()
to get all of them.

Change-Id: I6d8cacda4b62a6f6fdc77885149e7de8666e7200
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88196
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>