Resolves: tdf#130513 sfxlo-OptionalBox isn't in sfx
its in vcl not sfx, so under windows its not actually created by the
builder as the factory method isn't found in sfx. While under linux
the method is found anyway despite the wrong lib name.

seeing as its in vcl, keep it there. The factory method isn't needed
when its in vcl and rename the .ui name to VclOptionalBox

Change-Id: If279477d2e4cbb50d202d528c293b471134de30c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88279
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolanm@redhat.com>