reportdesign: Hide some internal commands from customization
In addition:

- .uno:BackColor was unused, and internally mapped to the same
id as .uno:DBBackgroundColor which is used and has a label. So
remove the former, and make the latter visible in the
customization dialog instead.

- .uno:DSBEditDoc is mapped to SID_EDITDOC, and handled by the
code, but has no label. On the other hand, the toolbar has an
.uno:EditDoc item, which has no mapping in reportdesign, but
happens to be mapped to SID_EDITDOC elsewhere. So assume this
was a mistake, and .uno:DSBEditDoc should be used instead of
.uno:EditDoc.

Change-Id: I9da2ed36167571627b32a4055b734847c3ea1e64
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88342
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>