Restore simple undo and redo buttons for the basctl module
Changed in commit c34edadf5bd3d1d9f3c9c056af28b8964d8f1ca0
("rework SvxUndoRedoControl to be a PopupWindowController"),
but the dropdowns do not work in that module.

These buttons used to be managed by SvxSimpleUndoRedoController,
but that's no longer the case. The reason is that
ToolBarManager::CreateControllers checks first for controllers
registered in Controller.xcu, and only if none found it checks
for sfx2 controllers. So SvxSimpleUndoRedoController by using
a sfx2-style registration, has no chance to be ever selected
for .uno:Undo or .uno:Redo.

This commit removes the unused controller, and restores the
previous behavior with the other controller.

Change-Id: Ia774195511e41ab11562856fe1cf2ec7f170710a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88606
Tested-by: Jenkins
Reviewed-by: Maxim Monastirsky <momonasmon@gmail.com>