Proper error check for fwrite in updater.cxx
A less-than-zero check for an unsigned value of type size_t
doesn't make sense, and:

If the return value of fwrite differs from the count parameter,
it means that a writing error prevented the function from completing.
In this case, the error indicator (ferror) will be set for the stream.

Change-Id: I344e608b2bc03a4a117fc5cca1acb8e26eda247a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88817
Tested-by: Jenkins
Reviewed-by: Muhammet Kara <muhammet.kara@collabora.com>