tdf#126785 qt5: Ignore external QEvent::ShortcutOverride
For some reason, an extra external QEvent::ShortcutOverride
event is triggered when a key is pressed and e.g. the
Orca screen reader is enabled, which would result in
duplicated input when handled as key input event as well.

Therefore, ignore those ShortcutOverride events that are
"spontaneous", i.e. "the event originated outside the
application (a system event)". [1]

[1] https://doc.qt.io/qt-5/qevent.html#spontaneous

Change-Id: Ibccacb69c31d4d7ebd512c704f0e4df978c5281f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/88852
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>