tdf#129553 sw_redlinehide: fix replace of paragraph breaks with $
The problem is that in the special selections that replace paragraph
breaks, the start and end node are actually different, hence have
different text frames.

(regression from 5e81b966778d82692b4763d892b457186a7f269d)

Practically the end content index is going to be 0 typically which
doesn't need mapping, but not sure if that's guaranteed anywhere.

Change-Id: I77c95c2f3e55edcc7aca6f4540cae90b3be5d79b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89050
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@cib.de>