tdf#130768 need to use mnOutOffX/mnOutOffY
in OutputDevice local stuff when want to get
to pixel coordiantes. These are not often used
local members of OutputDevice specially used
in fake-Windows as internal windows offset.
Thus there is the protected internal replacement
ImplGetDeviceTransformation() for the usually
used GetViewTransformation().
A very problematic thing - we should in principle
add this to GetViewTransformation() but I am sure
that this will lead to other problems - argh!

Change-Id: Ibedc3a7d6eb3f17c266082729872b81f607836a5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89259
Tested-by: Jenkins
Reviewed-by: Armin Le Grand <Armin.Le.Grand@me.com>