lok: remove sum and equal buttons from the formula bar
This patch removes sum and equal buttons from the formula bar in the
lok case. A tentative to limit that to the mobile case has been done
however the formula input bar is created before than the client view
so we are not able to use the LibreOfficeKit::isMobile check.

Change-Id: Icb820a0d77eb9295046997d2f3690d7f8b39680d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89453
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/93342
Tested-by: Jenkins