Keep selections selected after outline moves
and don't update content view until after move is complete.

In this patch a flag is used to not allow document changed notifications
to rebuild the content tree while doing outline moves. This flag also
prevents the content view from being update until after outline move is
complete. A condition is added in SwContentTree::Paint which prevents a
focus rectangle flash at the top of the list from SvTreeListBox::Paint
when the tree is empty.

Change-Id: I5691a532a6bb916c3152900eda3bc9333c24871e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89608
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>