graphic: remove preview when importing with GraphicFilter
The property "PreviewSizeHint" and the enum value ForPreview in
GraphicFilterImportFlags doesn't seem to be used from outside, so
let's remove it as it simplifies the PNG and JPEG import filters.

It is generally more recommended to load the bitmap file fully and
only then resize it to a smaller (preview) size.

Change-Id: I97f333686c25f7a7a4bdf6c6f0885154b515fd3d
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/89932
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>