cid#1460477 Result is not floating-point
warnings is about

aRange.getMinX()-nRadius/2

"
CID 1460477 (#1 of 1): Result is not floating-point (UNINTENDED_INTEGER_DIVISION)

integer_division: Dividing integer expressions 5 and 2, and then converting the
integer quotient to type double. Any remainder, or fractional part of the
quotient, is ignored.

To compute and use a non-integer quotient, change or cast either operand to
type double. If integer division is intended, consider indicating that by
casting the result to type int
"

Change-Id: I600d909c52b9e86f5193806d42cfd0421d8fc2cd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90132
Tested-by: Jenkins
Tested-by: Caolán McNamara <caolanm@redhat.com>
Reviewed-by: Caolán McNamara <caolanm@redhat.com>