basegfx: B2DTuple replace operator[] with "get" and Axis2D enum
opertaor[] was used for index access of of x and y variables
where 0 is x and 1 is y - similar like access to an array with 2
elements.

This comes in handy when you write generic code where the
algorithm is the same for x and y, but using index access operator
doesn't look clean and is potentially dangerous. We know we only
have 2 options (o and 1), but an index access allows for more.

A solution to this is to have a normal "get" method, with an enum
as parameter (Axis2D), which can only have values X or Y.

Change-Id: I3f98d0149214808a336f25599350a78436236827
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90133
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>