tdf#131153 Correct LongDateDaySeparator for fr_CH and it_CH
According to localeplanet [1] [2], in comparison to other software and people
living it these areas, the LongDateDaySeparator '.' for fr_CH and it_CH in
LibreOffice is not correct. It should be omitted.

[1] http://www.localeplanet.com/icu/it-CH/index.html
[2] http://www.localeplanet.com/icu/fr-CH/index.html

This means for the FormatElement index 22, 23, 24, 25, 26 should be the same in
fr_CH/fr_FR and it_CH/it_IT.

Change-Id: Ief4de0d8728c7a3bbcfac7f6200f37f2d2c647aa
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90427
Tested-by: Jenkins
Reviewed-by: Eike Rathke <erack@redhat.com>