Sidebar tdf#92768: Fix title & subtitle checkboxes on the ChartDeck
Turns out the 2nd part of tdf#92768 was not implemented yet, so it is
still necessary to remove the title/subtitle for good, otherwise we set
the "Visibility" property, but nothing actually reads that when
rendering.

So the titles now work as they did before the regression. An uncheck action
removes the title/subtitle, and a check action just makes the existing
title/subtitle visible or creates a new one if it doesn't exist

Change-Id: I23122d1be2d95af878b824ba194bd1aeed5f4f89
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90692
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Andras Timar <andras.timar@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/105737
Tested-by: Jenkins
Reviewed-by: Jan Holesovsky <kendy@collabora.com>