Extract testDisposedByDesktopTerminate from CppunitTest_sw_apitests
It is not guaranteed that SwXTextTable runs last during CppunitTest_sw_apitests,
and e.g. building with --enable-lto on Linux caused SwXTextField to be run after
SwXTextTable.  But then testDisposedByDesktopTerminate had already shut down all
the relevant global state, so the SwXTextField tests would fail.

TerminateTest is a faithful copy of the original SwXTextTable
testDisposedByDesktopTerminate behavior, but I do not know how much of that (if
anything at all) is actually test-worthy.  (For example, parts of
TerminateTest::init may not be relevant, or the original SwXTextTable
testDisposedByDesktopTerminate may even only have been there to clean up after
other tests and not to actually test anything by itself, cf.
20ed81fd42cf8079c11c2b18973b9504a679e49c "Make individual tests clean up after
themselves".)

Change-Id: Ib95b7fa6888f55a27236a779c453a6d04122a527
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90774
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <sbergman@redhat.com>