tdf#117448 Writer table: don't clip text on margins
Use area of paragraph margins to show top and bottom
of the clipped text at small fixed line height, like
MSO does. This results noticeable difference in tables,
where small fixed line height is a method to set narrow
table rows, but LibreOffice hid top and bottom of the
characters according to the line height which is smaller,
than the character height.

Note: PDF export has already showed top and bottom of the
characters in this case, and not in tables. But according
to the editing glitches (missing update of the clipped
area during typing), we don't want to extend this behaviour
for not table content, yet.

Change-Id: I7aff5405314948f301bfb71bf35cc1911e194f8b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/90798
Tested-by: Jenkins
Reviewed-by: László Németh <nemeth@numbertext.org>