no reason for the FontAntiAliasing setting being UNX-only (tdf#107744)
This dates back to the ancient times where no revision control history
dares, so no idea why this was so, but it doesn't make sense, as
at least Windows can disable AA for fonts too. And if some platform
really wants to hide the setting, it should be enough just to hide
the UI element, no reason for '#if defined( UNX )' all over the place.

Change-Id: I76fc9ad157358321904573079ace5a3fe556f86a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91041
Tested-by: Jenkins
Reviewed-by: Luboš Luňák <l.lunak@collabora.com>