basegfx: test B2DHomMatrix and Matrix together
This adds a test for B2DHomMatix and Matrix to test them together.
Ther reason for this is that we can remove Matrix in favour to
B2DHomMatrix, but first we need to make sure it is safe so by
checking that the operations in Matrix yield the same result as
with B2DHomMatrix.

Change-Id: Ia93c1d3c7b2b7ff257c35831ff31afad3e8f34ba
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/91312
Tested-by: Jenkins
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>